+++ lib/src/ui/features/admin/types/admin_error_code.dart
@@ -1,0 +1,4 @@
+enum AdminErrorCode {
+  missingTitleOrBody,
+  sendingMessageFailed,
+}